- Changed the type of reward.score from 'int' to 'float'  - FlyingPig is now a Collectible
Former-commit-id: 63d43dc50fb3fc979d8d6fd1084f919a922fd65b

+++ Assets/Game/Gameplay/Entities/Gelato.cs
@@ -15,8 +15,8 @@ public class Gelato : Entity {
 	}
 
 	protected override void OnRewardCreated() {
-		m_reward.score = (int)(m_reward.score * m_scoreMultiplier);
-		m_reward.coins = (int)(m_reward.coins * m_coinsMultiplier);
-		m_reward.xp = (int)(m_reward.xp * m_xpMultiplier);
+		m_reward.score = (m_reward.score * m_scoreMultiplier);
+		m_reward.coins = (m_reward.coins * m_coinsMultiplier);
+		m_reward.xp = (m_reward.xp * m_xpMultiplier);
 	}
 }
